# src/main/resources/api/openapi-spec.yaml
openapi: 3.0.3
...
info:
  title: Azul Bookstore API
  description: |
    A comprehensive REST API for managing books, authors, and genres in the Azul Bookstore system.

    ## Features
    - Full CRUD operations for Books, Authors, and Genres
    - Spring Boot-style pagination with flexible sorting
    - JWT-based authentication and authorization
    - Comprehensive error handling and validation
    - Relationship management between entities

    ## Authentication
    All protected endpoints require JWT bearer token authentication.
    Include the token in the Authorization header: `Authorization: Bearer <token>`

    ## Pagination
    List endpoints support pagination using query parameters:
    - `page`: Page number (0-based, default: 0)
    - `size`: Page size (default: 20, max: 100)
    - `sort`: Sort specification (e.g., `name,asc`, `title,desc`)

    ## Relationships
    - Books can have multiple authors and genres
    - Authors can be associated with multiple books
    - Genres are standalone entities used for categorization
  version: 1.0.0
  contact:
    name: Azul Bookstore API Support
    email: api-support@bookstore.azul.verglor.com
    url: https://api.bookstore.azul.verglor.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# Server configurations for different environments
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
    variables:
      version:
        default: v1
        enum: [v1]
...

...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Authorization header using the Bearer scheme.

        **Format:** `Authorization: Bearer <token>`

        Obtain tokens through the authentication endpoints (not included in this specification).

  # Common Parameters
  parameters:
...
    pageParam:
      name: page
      in: query
      description: Page number (0-based indexing)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
    sizeParam:
      name: size
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    sortParam:
      name: sort
      in: query
      description: |
        Sort specification compatible with Spring Boot's Sort functionality.

        **Format:** `field,direction` (can be repeated for multiple sort fields)

        **Direction options:** `asc`, `desc`, `ASC`, `DESC`

        **Examples:**
        - `?sort=price,asc` - Sort by price ascending
        - `?sort=title,desc` - Sort by title descending
        - `?sort=price,asc&sort=title,desc` - Sort by price ascending, then title descending

        **Note:** Multiple sort fields are supported by repeating the sort parameter and will be applied in order.
      required: false
      schema:
        type: array
        items:
          type: string
          minLength: 3
          maxLength: 50
      style: form
      explode: true
      example: name,asc

    # Path parameters
...
    authorIdParam:
      name: authorId
      in: path
      description: Unique identifier of the author
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 456
...
  schemas:
...
    BookResponse:
      type: object
      description: Book entity representing a publication in the bookstore
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the book
          example: 123
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Title of the book
          example: "The Great Gatsby"
        price:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          multipleOf: 0.01
          description: Price of the book in the store's currency
          example: 19.99
        authors:
          type: array
          description: List of authors associated with this book
          items:
            $ref: '#/components/schemas/AuthorResponse'
        genres:
          type: array
          description: List of genres associated with this book
          items:
            $ref: '#/components/schemas/GenreResponse'
      required:
        - title
        - price
    AuthorResponse:
      type: object
      description: Author entity representing a book author
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the author
          example: 456
        name:
          type: string
          minLength: 2
          maxLength: 50
          description: Full name of the author
          example: "F. Scott Fitzgerald"
      required:
        - name
    GenreResponse:
      type: object
      description: Genre entity representing a book category
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the genre
          example: 789
        name:
          type: string
          minLength: 2
          maxLength: 50
          description: Name of the genre
          example: "Fiction"
      required:
        - name

    # Pagination Schema
    PagedResponse:
      type: object
      description: Standard paginated response wrapper
      properties:
        content:
          type: array
          description: Array of items for the current page
          items:
            oneOf:
              - $ref: '#/components/schemas/BookResponse'
              - $ref: '#/components/schemas/AuthorResponse'
              - $ref: '#/components/schemas/GenreResponse'
        page:
          $ref: '#/components/schemas/PageInfo'
    PageInfo:
      type: object
      description: Pagination information
      properties:
        number:
          type: integer
          description: Current page number (0-based)
          example: 0
        size:
          type: integer
          description: Size of the current page
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          example: 8
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2023-12-01T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed for the request"
        validationErrors:
          type: array
          description: Detailed validation errors (when applicable)
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      description: Individual field validation error
      properties:
        field:
          type: string
          description: Name of the field that failed validation
          example: "title"
        rejectedValue:
          description: The value that was rejected
          example: ""
        message:
          type: string
          description: Validation error message
          example: "Book title is required"

    # Request Schemas (for creation and updates)
...
    AuthorRequest:
      type: object
      description: Request payload for creating or updating an author
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          description: Full name of the author
          example: "F. Scott Fitzgerald"
      required:
        - name
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2023-12-01T10:30:00Z"
            status: 400
            error: "Bad Request"
            message: "Validation failed"
            validationErrors:
              - field: "title"
                rejectedValue: ""
                message: "Book title is required"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2023-12-01T10:30:00Z"
            status: 401
            error: "Unauthorized"
            message: "JWT token is missing or invalid"
...
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2023-12-01T10:30:00Z"
            status: 404
            error: "Not Found"
            message: "Book with ID 999 not found"
    Conflict:
      description: Conflict - Resource already exists or constraint violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2023-12-01T10:30:00Z"
            status: 409
            error: "Conflict"
            message: "Genre with name 'Fiction' already exists"
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2023-12-01T10:30:00Z"
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"

# API Paths - CRUD operations for each entity
paths:
...
  /authors:
    # Get all authors with pagination
    get:
      summary: Get all authors
      description: Retrieve a paginated list of all authors
      operationId: getAllAuthors
      tags:
        - Authors
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
        - name: name
          in: query
          description: Search authors by name (case-insensitive partial match)
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 50
          example: "fitzgerald"
      x-spring-paginated: true
      responses:
        '200':
          description: Successfully retrieved authors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResponse'
              example:
                content:
                  - id: 1
                    name: "F. Scott Fitzgerald"
                page:
                  number: 0
                  size: 20
                  totalElements: 50
                  totalPages: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    # Create a new author
    post:
      summary: Create a new author
      description: Create a new author in the system
      operationId: createAuthor
      tags:
        - Authors
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
            example:
              name: "Harper Lee"
      responses:
        '201':
          description: Author created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                id: 2
                name: "Harper Lee"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Individual author operations
  /authors/{authorId}:
    # Get author by ID
    get:
      summary: Get author by ID
      description: Retrieve a specific author by their unique identifier
      operationId: getAuthorById
      tags:
        - Authors
      parameters:
        - $ref: '#/components/parameters/authorIdParam'
      responses:
        '200':
          description: Author found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                id: 1
                name: "F. Scott Fitzgerald"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    # Update author by ID
    put:
      summary: Update author by ID
      description: Update an existing author's information
      operationId: updateAuthor
      tags:
        - Authors
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/authorIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
            example:
              name: "Francis Scott Fitzgerald"
      responses:
        '200':
          description: Author updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                id: 1
                name: "Francis Scott Fitzgerald"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    # Delete author by ID
    delete:
      summary: Delete author by ID
      description: Permanently delete an author from the system
      operationId: deleteAuthor
      tags:
        - Authors
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/authorIdParam'
      responses:
        '204':
          description: Author deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Genre endpoints
...